backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.7
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.11
alembic==1.11.1
python-multipart==0.0.6
bcrypt==4.0.1
